# perf: remove now needless canonicalization getting closest package.json