This is no longer required because we now store everything in the file system for deno compile instead of in an eszip.

Before:

```
Timer precision: 41 ns
bench          fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ resolution  217.7 µs      │ 715.5 µs      │ 227.2 µs      │ 233.4 µs      │ 1000    │ 1000
```

After:

```
Timer precision: 41 ns
bench          fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ resolution  191.1 µs      │ 768.7 µs      │ 198.2 µs      │ 205.4 µs      │ 1000    │ 1000
```

(My system is noisy, not doing a proper bench, but it removes some fs operations so going to be faster)